Reverts 2388aae4 [NARY-REASSOCIATE] Support reassociation of min/max

+++ lld/MachO/SymbolTable.cpp
@@ -159,9 +159,9 @@ Symbol *SymbolTable::addLazy(StringRef name, ArchiveFile *file,
 }
 
 Defined *SymbolTable::addSynthetic(StringRef name, InputSection *isec,
-                                   uint32_t value, bool isPrivateExtern,
+                                   uint64_t value, bool isPrivateExtern,
                                    bool includeInSymtab) {
-  Defined *s = addDefined(name, nullptr, isec, value, /*size*/ 0,
+  Defined *s = addDefined(name, nullptr, isec, value, /*size=*/0,
                           /*isWeakDef=*/false, isPrivateExtern);
   s->includeInSymtab = includeInSymtab;
   return s;
